requirements.txt
openai
faiss-cpu
chromadb
tiktoken
python-dotenv
langchain-openai
langchain-community
unstructured
markdown
docx2txt
streamlit